strainjedi: move get_rics to rics submodule
While we're here, also use explicit namespacing of ASE units for
clarity.
